fdsf

+++ src/prog2/model/CamiTerra.java
@@ -2,7 +2,7 @@ package prog2.model;
 
 public class CamiTerra extends AccesTerra{
 
-    public CamiTerra(String nomacc, boolean estat, boolean accessibilitat, LlistaAllotjaments llistaAllotjaments, double longitud) {
+    public CamiTerra(String nomacc, boolean estat, boolean accessibilitat, LlistaAllotjaments llistaAllotjaments, float longitud) {
         super(nomacc, estat, longitud);
     }
